+++ sql/innodb_priv.h
@@ -22,7 +22,7 @@
 
 class THD;
 
-int get_quote_char_for_identifier(THD *thd, const char *name, uint length);
+int get_quote_char_for_identifier(THD *thd, const char *name, size_t length);
 bool schema_table_store_record(THD *thd, TABLE *table);
 void localtime_to_TIME(MYSQL_TIME *to, struct tm *from);
 uint strconvert(CHARSET_INFO *from_cs, const char *from, uint from_length,
